requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Werkzeug==2.3.7
python-dotenv==1.0.0
authlib==1.2.1
requests==2.31.0